Treino de Spring melhorando o método de remover

+++ src/main/java/com/zap/lojazap/api/controller/CozinhaController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -85,11 +86,29 @@ public class CozinhaController {
 //		return ResponseEntity.ok(cozinhaAtual);
 //	}
 	
-	
 	@DeleteMapping("/{id}")
-	public void remover(@PathVariable Long  id) {
+	public ResponseEntity<CozinhaEntity> remover(@PathVariable Long  id) {
 		CozinhaEntity cozinha = cozinhaRepository.porId(id);
-		cozinhaRepository.remover(cozinha);
+		try {
+			if(cozinha != null) {
+				cozinhaRepository.remover(cozinha);
+				
+				return ResponseEntity.noContent().build();
+			}
+			
+			return ResponseEntity.notFound().build();
+			
+		} catch (DataIntegrityViolationException e) {
+			return ResponseEntity.badRequest().build(); //poderia ser tbm o CONFLICT
+		}
+
 	}
+	
+	
+//	@DeleteMapping("/{id}")
+//	public void remover(@PathVariable Long  id) {
+//		CozinhaEntity cozinha = cozinhaRepository.porId(id);
+//		cozinhaRepository.remover(cozinha);
+//	}
 
 }
